Change celerity type from int to float

+++ backend/models/project.js
@@ -6,7 +6,7 @@ module.exports = (sequelize, DataTypes) => {
       name: DataTypes.STRING,
       thirdPartyType: DataTypes.ENUM('NONE', 'TRELLO', 'JIRA'),
       thirdPartyId: { type: DataTypes.STRING, unique: true },
-      celerity: { type: DataTypes.INTEGER, defaultValue: 6 },
+      celerity: { type: DataTypes.FLOAT, defaultValue: 6.0 },
       dailyDevelopmentTime: { type: DataTypes.INTEGER, defaultValue: 21600000 }, // 6 * 3600 * 1000
     },
     {},
